requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-tailwind==0.5.0
cron_descriptor==2.0.5
distro==1.9.0
Django==4.2.11
django-browser-reload==1.18.0
django-celery-beat==2.5.0
django-crispy-forms==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kombu==5.5.4
openai==1.106.0
packaging==25.0
pillow==10.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==7.12.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
